# openapi.yaml
openapi: 3.1.0
info:
  title: ChessDesk Backend API
  version: 0.1.0
  description: >-
    API contract for the ChessDesk frontend. Authenticated requests use the
    session cookie set by the authentication endpoints.
servers:
  - url: /api/v1
    description: Backend API base path
...
security:
  - sessionCookie: []
paths:
...
  /clients:
    get:
      tags: [Clients]
      operationId: getClients
      summary: List clients
      description: Returns clients owned by the specified coach, enriched with details, next session, and outstanding amount.
      parameters:
        - $ref: '#/components/parameters/CoachIdQuery'
      responses:
        '200':
          description: Matching clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClientWithDetails'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags: [Clients]
      operationId: createClient
      summary: Create a client
      description: Creates either an individual student client or a school client. Include the matching details object for the selected client type.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientCreateRequest'
            examples:
              individual:
                value:
                  coach_id: coach-001
                  client_type: individual
                  display_name: Amahle Dlamini
                  email: parent.dlamini@example.com
                  whatsapp: '+27 73 112 3344'
                  preferred_communication: whatsapp
                  notifications_enabled: true
                  notes: Preparing for U14 provincials.
                  active: true
                  individual_details:
                    student_name: Amahle Dlamini
                    school_name: Rosebank College Prep
                    parent_name: Zanele Dlamini
              school:
                value:
                  coach_id: coach-001
                  client_type: school
                  display_name: Greenfields Primary
                  email: admin@greenfieldsprimary.example.com
                  whatsapp: '+27 11 456 7890'
                  preferred_communication: email
                  notifications_enabled: true
                  notes: Wednesday afternoons.
                  active: true
                  school_details:
                    school_name: Greenfields Primary
                    contact_person: Mrs. Karen Botha
                    learner_range: 20-30
      responses:
        '201':
          description: Created client (without the computed/enriched fields)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /clients/{clientId}:
    parameters:
      - $ref: '#/components/parameters/ClientId'

    get:
      tags: [Clients]
      operationId: getClient
      summary: Get one client
      responses:
        '200':
          description: Client with details and computed values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientWithDetails'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Clients]
      operationId: updateClient
      summary: Update a client
      description: Updates only the supplied mutable client fields. The client id and owning coach are immutable.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientUpdateRequest'
      responses:
        '200':
          description: Updated client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Clients]
      operationId: deleteClient
      summary: Delete a client
      responses:
        '204':
          description: Client deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /invoices:
    get:
      tags: [Invoices]
      operationId: getInvoices
      summary: List invoices
      parameters:
        - $ref: '#/components/parameters/CoachIdQuery'
      responses:
        '200':
          description: Matching invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Invoice'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags: [Invoices]
      operationId: createInvoice
      summary: Create an invoice
      description: session_ids contains zero or more sessions attached to the invoice.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoiceCreateRequest'
      responses:
        '201':
          description: Created invoice
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /invoices/{invoiceId}:
    parameters:
      - $ref: '#/components/parameters/InvoiceId'

    patch:
      tags: [Invoices]
      operationId: updateInvoice
      summary: Update an invoice
      description: >-
        Used by updateInvoice and payment status actions in the frontend.
        markPaid sends status=paid, paid_date, payment_method, and
        payment_reference. Reversing a mistaken paid status sends status=unpaid,
        null paid_date and payment_method, and an empty payment_reference.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoiceUpdateRequest'
            examples:
              markPaid:
                value:
                  status: paid
                  paid_date: '2026-09-07'
                  payment_method: eft
                  payment_reference: EFT-20260907-001
              markUnpaid:
                value:
                  status: unpaid
                  paid_date: null
                  payment_method: null
                  payment_reference: ''
      responses:
        '200':
          description: Updated invoice
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Invoices]
      operationId: deleteInvoice
      summary: Delete an invoice
      responses:
        '204':
          description: Invoice deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /invoices/{invoiceId}/sessions:
    get:
      tags: [Invoices]
      operationId: getSessionsForInvoice
      summary: List session ids attached to an invoice
      parameters:
        - $ref: '#/components/parameters/InvoiceId'
      responses:
        '200':
          description: Attached session ids
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: chessdesk_session
      description: HttpOnly session cookie established by login or signup.
...
  parameters:
...
    CoachIdQuery:
      name: coach_id
      in: query
      required: true
      description: Coach whose records should be returned. The authenticated user must own this id.
      schema:
        type: string
      example: coach-001
    ClientId:
      name: clientId
      in: path
      required: true
      schema:
        type: string
      example: client-001
...
    InvoiceId:
      name: invoiceId
      in: path
      required: true
      schema:
        type: string
      example: inv-001
...
  responses:
    BadRequest:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication is required or the session is invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: The authenticated coach cannot access this resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
        code:
          type: string
        field_errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
...
    ClientType:
      type: string
      enum: [individual, school]
    CommunicationPreference:
      type: string
      enum: [whatsapp, email, both]
    LearnerRange:
      type: string
      enum: ['1-10', '10-20', '20-30', '30-40', '40+']
    Client:
      type: object
      required:
        - id
        - coach_id
        - client_type
        - display_name
        - email
        - whatsapp
        - preferred_communication
        - notifications_enabled
        - notes
        - active
      properties:
        id:
          type: string
          readOnly: true
        coach_id:
          type: string
          readOnly: true
        client_type:
          $ref: '#/components/schemas/ClientType'
        display_name:
          type: string
        email:
          type: string
          format: email
        whatsapp:
          type: string
        preferred_communication:
          $ref: '#/components/schemas/CommunicationPreference'
        notifications_enabled:
          type: boolean
        notes:
          type: string
        active:
          type: boolean
    IndividualStudentDetails:
      type: object
      required: [client_id, student_name, school_name, parent_name]
      properties:
        client_id:
          type: string
          readOnly: true
        student_name:
          type: string
        school_name:
          type: string
          description: Optional school name; the frontend represents an omitted value as an empty string.
        parent_name:
          type: string
    SchoolDetails:
      type: object
      required: [client_id, school_name, contact_person, learner_range]
      properties:
        client_id:
          type: string
          readOnly: true
        school_name:
          type: string
        contact_person:
          type: string
        learner_range:
          $ref: '#/components/schemas/LearnerRange'
    ClientCreateRequest:
      type: object
      required:
        - coach_id
        - client_type
        - display_name
        - email
        - whatsapp
        - preferred_communication
        - notifications_enabled
        - notes
        - active
      properties:
        coach_id:
          type: string
        client_type:
          $ref: '#/components/schemas/ClientType'
        display_name:
          type: string
        email:
          type: string
          format: email
        whatsapp:
          type: string
        preferred_communication:
          $ref: '#/components/schemas/CommunicationPreference'
        notifications_enabled:
          type: boolean
        notes:
          type: string
        active:
          type: boolean
        individual_details:
          description: Required by the frontend when client_type is individual.
          $ref: '#/components/schemas/IndividualStudentDetailsCreate'
        school_details:
          description: Required by the frontend when client_type is school.
          $ref: '#/components/schemas/SchoolDetailsCreate'
    ClientUpdateRequest:
      type: object
      minProperties: 1
      properties:
        client_type:
          $ref: '#/components/schemas/ClientType'
        display_name:
          type: string
        email:
          type: string
          format: email
        whatsapp:
          type: string
        preferred_communication:
          $ref: '#/components/schemas/CommunicationPreference'
        notifications_enabled:
          type: boolean
        notes:
          type: string
        active:
          type: boolean
    IndividualStudentDetailsCreate:
      type: object
      required: [student_name, school_name, parent_name]
      properties:
        student_name:
          type: string
        school_name:
          type: string
        parent_name:
          type: string
    SchoolDetailsCreate:
      type: object
      required: [school_name, contact_person, learner_range]
      properties:
        school_name:
          type: string
        contact_person:
          type: string
        learner_range:
          $ref: '#/components/schemas/LearnerRange'
    ClientWithDetails:
      allOf:
        - $ref: '#/components/schemas/Client'
        - type: object
          required: [outstanding_amount]
          properties:
            individual_details:
              $ref: '#/components/schemas/IndividualStudentDetails'
            school_details:
              $ref: '#/components/schemas/SchoolDetails'
            upcoming_session:
              type: [string, 'null']
              description: Date and time formatted as YYYY-MM-DD HH:mm, or null when none is upcoming.
            outstanding_amount:
              type: number
              minimum: 0
...
    PaymentMethod:
      type: string
      enum: [eft, cash, other]
    InvoiceStatus:
      type: string
      enum: [unpaid, paid]
    Invoice:
      type: object
      required:
        - id
        - coach_id
        - client_id
        - invoice_date
        - due_date
        - amount
        - description
        - status
        - paid_date
        - payment_method
        - payment_reference
        - notes
      properties:
        id:
          type: string
          readOnly: true
        coach_id:
          type: string
          readOnly: true
        client_id:
          type: string
        invoice_date:
          type: string
          format: date
        due_date:
          type: string
          format: date
        amount:
          type: number
          minimum: 0
        description:
          type: string
        status:
          $ref: '#/components/schemas/InvoiceStatus'
        paid_date:
          type: [string, 'null']
          format: date
        payment_method:
          anyOf:
            - $ref: '#/components/schemas/PaymentMethod'
            - type: 'null'
        payment_reference:
          type: string
        notes:
          type: string
    InvoiceCreateRequest:
      type: object
      required:
        - coach_id
        - client_id
        - invoice_date
        - due_date
        - amount
        - description
        - status
        - paid_date
        - payment_method
        - payment_reference
        - notes
        - session_ids
      properties:
        coach_id:
          type: string
        client_id:
          type: string
        invoice_date:
          type: string
          format: date
        due_date:
          type: string
          format: date
        amount:
          type: number
          minimum: 0
        description:
          type: string
        status:
          $ref: '#/components/schemas/InvoiceStatus'
        paid_date:
          type: [string, 'null']
          format: date
        payment_method:
          anyOf:
            - $ref: '#/components/schemas/PaymentMethod'
            - type: 'null'
        payment_reference:
          type: string
        notes:
          type: string
        session_ids:
          type: array
          items:
            type: string
    InvoiceUpdateRequest:
      type: object
      minProperties: 1
      properties:
        client_id:
          type: string
        invoice_date:
          type: string
          format: date
        due_date:
          type: string
          format: date
        amount:
          type: number
          minimum: 0
        description:
          type: string
        status:
          $ref: '#/components/schemas/InvoiceStatus'
        paid_date:
          type: [string, 'null']
          format: date
        payment_method:
          anyOf:
            - $ref: '#/components/schemas/PaymentMethod'
            - type: 'null'
        payment_reference:
          type: string
        notes:
          type: string
        session_ids:
          type: array
          items:
            type: string
...